requirements.txt
python-dotenv
requests
huggingface_hub
langchain-core
langchain-community
duckduckgo-search
